Get Immundata internal schema field names

Returns the standardized field names used across Immundata objects and processing functions, as defined in IMD_GLOBALS$schema. These include column names for cell ids or barcodes, receptors, repertoires, and related metadata.

Usage

imd_schema(key = NULL)

imd_schema_sym(key = NULL)

imd_meta_schema()

imd_files()

imd_rename_cols(format = "default")

imd_drop_cols(format = "airr")

imd_repertoire_schema(format = "airr")

imd_receptor_features(schema)

imd_receptor_chains(schema)

Arguments

key

Character which field to return.

format

Character what format to load - "airr" or "10x".

schema

Receptor schema from make_receptor_schema().